backend/requirements.txt
cryptography==44.0.1
email-validator==2.2.0
fastapi==0.115.12
passlib==1.7.4
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
PyJWT==2.10.1
slowapi==0.1.9
sqlalchemy==2.0.40
uvicorn[standard]==0.34.0

